feat(33-04): x64 configs across boot-path lib tier + width fixes; plugin ProjectConfigs
- ~56 engine StaticLibrary .vcxprojs + zlib/ui/udplibrary: add Debug|x64 + Release|x64
  importing x64-platform.props, isolated \x64\ OutDir; Win32 blocks untouched (additive)
- 3 D3D9 plugin .vcxprojs (Direct3d9/_ffp/_vsps): x64 ProjectConfiguration + Configuration
  PropertyGroups only (reviews fix #9) -- NO Link/MachineX64 (Plan 05 owns those)
- x64 width fixes (BITS-02, surfaced by /we4311/4312 + SDK header width on x64):
  * sharedNetwork Sock.h + udplibrary UdpLibrary.hpp: SOCKET shadow -> winsock2/UINT_PTR
    (C2371 'different basic types'); guarded so Win32 ABI unchanged
  * crypto misc.cpp/h: alignment-test pointer casts unsigned int -> size_t
  * ui UITableModel: pointer-as-int handle via intptr_t; UILoader ptrdiff_t narrow;
    UITabbedPane _asm nop -> __noop (x64-illegal asm)
  * clientGraphics BinkTreeFileIO: file handle U32 -> UINTa (pointer-width round-trip)
  * clientGame WaterTestAppearance: sort-key pointer cast via intptr_t
  * clientAudio Audio.cpp: Miles AIL_file_*_callback ABI U32 -> UINTa (matches Mss.h)
  * blat.vcxproj: x64 include dirs + GENSOCK_STATIC_LINK
- Full boot-path lib tier compiles /p:Platform=x64 clean (62/62, 0 error C); Win32 non-regression verified

+++ src/engine/client/library/clientGame/src/shared/test/WaterTestAppearance.cpp
@@ -94,7 +94,7 @@ float WaterTestAppearance::LocalShaderPrimitive::getDepthSquaredSortKey () const
 
 int WaterTestAppearance::LocalShaderPrimitive::getVertexBufferSortKey () const
 {
-	return reinterpret_cast<int> (&m_appearance.getVertexBuffer ());
+	return static_cast<int>(reinterpret_cast<intptr_t> (&m_appearance.getVertexBuffer ()));	// x64: sort key = low-32 of the pointer (Phase 33 X64-01, BITS-02)
 }
 
 // ----------------------------------------------------------------------
